Added 2nd case of Nice connectors, and extracted superClass
